mcb_validate::ast

Trait Callback

pub trait Callback {
    type Res;
    type Cfg;

    // Required method
    fn call<T>(cfg: Self::Cfg, parser: &T) -> Self::Res
       where T: ParserTrait;
}

A trait for callback functions.

Allows to call a private library function, getting as result its output value.

Required Associated Types

type Res

The output type returned by the callee

type Cfg

The input type used by the caller to pass the arguments to the callee

Required Methods

fn call<T>(cfg: Self::Cfg, parser: &T) -> Self::Res
where T: ParserTrait,

Calls a function inside the library and returns its value

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.