mcb_validate

Macro define_validator

macro_rules! define_validator {
    (
        name: $name_str:expr,
        description: $desc_str:literal,

        $(#[$struct_meta:meta])*
        $struct_vis:vis struct $validator_name:ident {
            $(
                $(#[$field_meta:meta])*
                $field_vis:vis $field:ident : $field_ty:ty
            ),* $(,)?
        }

        violations: $severity_mode:ident, $category:expr,
        $(#[$enum_meta:meta])*
        $enum_vis:vis enum $enum_name:ident {
            $($violation_body:tt)*
        }

        checks: [
            $( $check_name:ident => $check_expr:expr ),* $(,)?
        ]

        $(, enabled = $enabled_expr:expr )?
        $(,)?
    ) => { ... };
}

Defines a complete validator including its struct, violation enum, and traits.

This unified macro replaces the need for separate define_violations!, pub struct, and impl Validator blocks.

It generates:

The violation enum (via define_violations!)

The validator struct

A validate_all(&self) convenience method

The Validator trait implementation

Note: You still need to implement new() and/or with_config() yourself in a separate impl block, as configuration loading varies between validators.